portable: 修复 32 位 time_t 平台毫秒时钟溢出
ts.tv_sec * 1000 在 time_t 宽度内先行计算再扩展到 uint64。在 32 位
time_t 平台（CI 的 linux-arm、android-armeabi-v7a）上，系统运行约
24.8 天后该乘法有符号溢出，导致所有串口超时逻辑失效。先转 uint64
再相乘。

+++ libportable/src/time_monotonic.c
@@ -8,7 +8,7 @@ time_monotonic()
 {
 	struct timespec ts;
 	clock_gettime(CLOCK_MONOTONIC, &ts);
-	return ts.tv_sec * 1000 + ts.tv_nsec / 1000 / 1000;
+	return (uint64_t)ts.tv_sec * 1000 + ts.tv_nsec / 1000 / 1000;
 }
 
 #endif
